Move socket_target method into lib
This patch moves the method to library and makes a library call from
main. It also disallows user to call cli without --target.
